RAG/项目四_高级检索优化RAG/requirements.txt
# 基础依赖
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymupdf==1.23.8
markdown==3.5.1
python-dotenv==1.0.0
openai==1.3.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3

# 高级检索优化特定依赖
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
protobuf==4.25.1
rank-bm25==0.2.2
scikit-learn==1.3.0
